+++ stats.c
@@ -7,7 +7,7 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -15,7 +15,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * Foundation, Inc., 51 Franklin Street, Suite 500, Boston, MA  02110-1335  USA
  */
 
 #define _GNU_SOURCE
@@ -45,8 +45,8 @@ void set_stat_prefix(const char* prefix)
 	strncpy(stat_prefix, prefix, sizeof(stat_prefix)-1);
 }
 
-static void scale_mult_unit(char *unit, int unitsize, 
-		       const char *baseunit, 
+static void scale_mult_unit(char *unit, int unitsize,
+		       const char *baseunit,
 		       double *value_min,
 		       double *value_avg,
 		       double *value_max)
@@ -54,7 +54,7 @@ static void scale_mult_unit(char *unit, int unitsize,
 	int mult = 0;
 	char multchar[] = "KMGTPE";
 
-	while ((*value_min >= 1024.0) && (*value_avg >= 1024.0) && 
+	while ((*value_min >= 1024.0) && (*value_avg >= 1024.0) &&
 	       (*value_max >= 1024.0) && (mult < sizeof(multchar))) {
 		mult++;
 		*value_min = *value_min / 1024.0;
@@ -105,7 +105,8 @@ char *dump_stat_counter(char *buf, int size,
 		       const char *msg, uint64_t value)
 {
 	buf[size-1] = 0;
-	snprintf(buf, size-1, "%s%s: %llu", stat_prefix, msg, value);
+	snprintf(buf, size-1, "%s%s: %llu", stat_prefix, msg,
+		 (unsigned long long) value);
 
 	return buf;
 }
@@ -120,14 +121,15 @@ char *dump_stat_stat(char *buf, int size,
 
 	buf[size-1] = 0;
 	snprintf(buf, size-1, "%s%s: (min=%llu; avg=%.3f; max=%llu)%s",
-		 stat_prefix, msg, stat->min, avg, stat->max, unit);
+		 stat_prefix, msg, (unsigned long long) stat->min, avg,
+		 (unsigned long long) stat->max, unit);
 
 	return buf;
 }
 
 char *dump_stat_bw(char *buf, int size,
-		  const char *msg, const char *unit, 
-		  struct rng_stat *stat, 
+		  const char *msg, const char *unit,
+		  struct rng_stat *stat,
 		  uint64_t blocksize)
 {
 	char unitscaled[20];
